requirements.txt
faker>=8.0.0
camoufox>=1.0.0
httpx>=0.25.0
pyyaml>=6.0.0
aiohttp>=3.9.0

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
gradio>=4.0.0

sqlalchemy>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
